Inode and Directory Structure added
