--- ftype.h
@@ -1,12 +1,0 @@
-/*Header file which contains the mappings for file types for our filesystem
- * Ravi Shreyas Anupindi
- * Rishabh Dutt
- * Rajat Nigam
-*/
-
-#ifndef FTPYE_H_
-#define FTYPE_H_
-
-
-
-#endif
